requirements.txt
# Python dependencies for AI Signal Engine and Advanced Crawlers
openai==1.6.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.25.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-telegram-bot==20.7
aiohttp==3.9.1
schedule==1.2.0
tweepy==4.14.0
discord.py==2.3.2

